@@ -10,32 +10,31 @@
     "clean": "rm -rf ./build && mkdir build"
   },
   "engines": {
-    "node": ">=12.11.1"
+    "node": ">=21.1.0"
   },
   "author": "",
   "license": "ISC",
   "dependencies": {
-    "@babel/cli": "^7.13.16",
-    "@babel/core": "^7.14.2",
-    "@babel/node": "^7.14.2",
-    "@babel/preset-env": "^7.14.2",
-    "@netlify/functions": "^2.6.3",
+    "@babel/cli": "^7.24.5",
+    "@babel/core": "7.24.5",
+    "@babel/node": "^7.23.9",
+    "@babel/preset-env": "^7.24.5",
+    "@netlify/functions": "^2.7.0",
     "@types/express": "^4.17.21",
     "bcryptjs": "^2.4.3",
     "body-parser": "^1.19.0",
-    "cors": "^2.8.5",
     "express": "^4.19.2",
-    "express-validator": "^6.11.1",
-    "jsonwebtoken": "^9.0.2",
-    "mongoose": "^5.12.9",
+    "express-validator": "^7.0.1",
+    "jsonwebtoken": "9.0.2",
+    "mongoose": "8.3.5",
     "morgan": "^1.10.0",
     "serverless-http": "^3.2.0"
   },
   "devDependencies": {
-    "eslint": "^7.26.0",
-    "eslint-config-prettier": "^8.3.0",
-    "eslint-plugin-prettier": "^3.4.0",
+    "eslint": "^9.2.0",
+    "eslint-config-prettier": "^9.1.0",
+    "eslint-plugin-prettier": "^5.1.3",
     "nodemon": "^3.1.0",
-    "prettier": "2.2.1"
+    "prettier": "^3.2.5"
   }
 }
