@@ -1,15 +1,15 @@
-const express = require('express');
-const errorHandler = require('./modules/core/errorHandler');
-const logger = require('./modules/core/logger');
-const parseResponse = require('./modules/core/parseResponse');
-const ignoreFavicon = require('./modules/core/ignoreFavicon');
-const cors = require('./modules/core/cors');
-const routes = require('./modules/core/routes');
-const dbConnect = require('./modules/core/db');
+import express from 'express';
+
+import errorHandler from './modules/core/errorHandler';
+import logger from './modules/core/logger';
+import parseResponse from './modules/core/parseResponse';
+import ignoreFavicon from './modules/core/ignoreFavicon';
+import cors from './modules/core/cors';
+import routes from './modules/core/routes';
+import dbConnect from './modules/core/db';
 
 const app = express();
 const PORT = process.env.PORT || 8000;
-
 dbConnect();
 logger(app);
 parseResponse(app);
