@@ -1,5 +1,14 @@
-const cors = require('cors');
-
-module.exports = (app) => {
-  app.use(cors());
-};
+export default function cors(app) {
+  app.use((req, res, next) => {
+    res.header('Access-Control-Allow-Origin', '*');
+    res.header(
+      'Access-Control-Allow-Headers',
+      'Origin, X-Requested-With, Content-Type, Accept, Authorization',
+    );
+    if (req.method === 'OPTIONS') {
+      res.header('Access-Control-Allow-Methods', 'PUT, POST, PATH, DELETE,GET');
+      return res.status(200).json({});
+    }
+    next();
+  });
+}
