@@ -1,4 +1,4 @@
-const mongoose = require('mongoose');
+import mongoose from 'mongoose';
 
 const {
   NODE_ENV,
@@ -9,22 +9,18 @@ const {
 let connectionString;
 
 switch (NODE_ENV) {
-  case 'local':
-    connectionString = MONGO_CONNECTION_STRING_LOCAL;
-    break;
-  default:
-    connectionString = MONGO_CONNECTION_STRING;
+case 'local':
+  connectionString = MONGO_CONNECTION_STRING_LOCAL;
+  break;
+default:
+  connectionString = MONGO_CONNECTION_STRING;
 }
 
 const options = {
-  useUnifiedTopology: true,
-  useCreateIndex: true,
-  useNewUrlParser: true,
   autoIndex: false,
-  useFindAndModify: false,
 };
 
-function dbConnect() {
+export default function dbConnect() {
   mongoose
     .connect(connectionString, options)
     .then(() => console.log('Mongo connected'))
@@ -32,9 +28,7 @@ function dbConnect() {
 
   const db = mongoose.connection;
   db.on('error', console.error.bind(console, 'connection error:'));
-  db.once('open', function () {
+  db.once('open', function() {
     console.log('CONNECTED');
   });
 }
-
-module.exports = dbConnect;
