@@ -1,6 +1,7 @@
-module.exports = (app) => {
-  app.use((err, req, res, next) => {
-    console.error(err.stack);
-    res.status(500).send('Something broke!');
-  });
-};
+function apiNotFound(req, res) {
+  res.status(400).json('API not found');
+}
+
+export default function errorHandler(app) {
+  app.use(apiNotFound);
+}
