@@ -1,9 +1,3 @@
-module.exports = (app) => {
-  app.use((req, res, next) => {
-    if (req.originalUrl === '/favicon.ico') {
-      res.status(204).json({ noContent: true });
-    } else {
-      next();
-    }
-  });
-};
+export default function ignoreFavicon(app) {
+  app.get('/favicon.ico', (req, res) => res.status(204));
+}
