@@ -1,6 +1,6 @@
-const bodyParser = require('body-parser');
+import bodyParser from 'body-parser';
 
-module.exports = function parseResponse(app) {
+export default function parseResponse(app) {
   app.use(bodyParser.urlencoded({ extended: false }));
   app.use(bodyParser.json());
-};
+}
