@@ -1,9 +1,9 @@
-const infoRouter = require('../info/Routes');
-const drawRouter = require('../draw/Routes');
-// const home = require('../home/home');
+import infoRouter from '../info/Routes';
+import drawRouter from '../draw/Routes';
+// import home from '../home/home';
 
-module.exports = function routes(app) {
-  // app.use('/', home);
+export default function routes(app) {
+  //app.use('/', home);
   app.use('/info', infoRouter);
   app.use('/draw', drawRouter);
-};
+}
